+++ packages/control/algorithm/additional_current_test.py
@@ -5,7 +5,8 @@ from control.algorithm import additional_current
 from control.chargepoint.chargepoint import Chargepoint, ChargepointData
 from control.chargepoint.chargepoint_data import Set
 from control.chargepoint.control_parameter import ControlParameter
-from control.ev import ChargeTemplate, Ev
+from control.ev.charge_template import ChargeTemplate
+from control.ev.ev import Ev
 from control.loadmanagement import LimitingValue
