+++ packages/control/algorithm/common_test.py
@@ -6,7 +6,7 @@ import pytest
 from control import data
 from control.algorithm import common
 from control.chargepoint.chargepoint import Chargepoint
-from control.ev import Ev
+from control.ev.ev import Ev
 from control.counter import Counter
 from control.counter_all import CounterAll
