+++ packages/control/algorithm/surplus_controlled_test.py
@@ -11,7 +11,8 @@ from control.chargepoint.chargepoint import Chargepoint, ChargepointData
 from control.chargepoint.chargepoint_data import Get, Set
 from control.chargepoint.chargepoint_template import CpTemplate
 from control.chargepoint.control_parameter import ControlParameter
-from control.ev import ChargeTemplate, Ev
+from control.ev.charge_template import ChargeTemplate
+from control.ev.ev import Ev
 
 
 @pytest.fixture(autouse=True)
