+++ packages/control/auto_phase_switch_test.py
@@ -8,7 +8,7 @@ from control.counter import Counter, CounterData, Set
 from control.pv_all import PvAll
 from control.bat_all import BatAll
 from control.general import General
-from control.ev import Ev
+from control.ev.ev import Ev
 from control import data
 from control.chargepoint.chargepoint_state import ChargepointState
